Add support for deleting images
