+++ app/Http/Livewire/Images/PaginatedImageList.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace App\Http\Livewire\Images;
+
+use App\Models\Image;
+use Livewire\Component;
+use Livewire\WithPagination;
+
+class PaginatedImageList extends Component
+{
+    use WithPagination;
+
+    /**
+     * The events the component should listen for.
+     *
+     * @var array
+     */
+    protected $listeners = ['imageDeleted' => '$refresh'];
+
+    /**
+     * Render the component.
+     *
+     * @return \Illuminate\View\View
+     */
+    public function render()
+    {
+        return view('images.paginated-image-list', [
+            'images' => Image::latest()->paginate(24),
+        ]);
+    }
+}
