+++ resources/views/images/index.blade.php
@@ -7,7 +7,7 @@
 
     <div class="py-12">
         <div class="max-w-7xl mx-auto sm:px-6 lg:px-8">
-            @livewire('paginated-image-list')
+            @livewire('images.paginated-image-list')
         </div>
     </div>
 </x-app-layout>
